+++ internal/syslog/facility_string.go
