+++ internal/syslog/message.go
