+++ internal/syslog/severity_string.go
